Created Shopping Bag component
